# docs/api_documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HydraX BITTEN Trading API",
    "description": "Production-ready trading signal API with rate limiting and authentication",
    "version": "1.0.0",
    "contact": {
      "name": "HydraX Support",
      "url": "https://joinbitten.com"
    }
  },
  "servers": [
    {
      "url": "https://joinbitten.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8888",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/signals": {
      "get": {
        "summary": "API endpoint with lazy loading",
        "description": "Function: api_signals (from webapp_server_optimized.py)",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/fire": {
      "post": {
        "summary": "Execute a trade (fire button)",
        "description": "Function: fire_trade (from web_app.py)",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}